// src/types/common/file.ts
export enum MIMES {
  png = "image/png",
  jpg = "image/jpeg",
  pdf = "application/pdf",
  ppt = "application/vnd.ms-powerpoint",
  pptx = "application/vnd.openxmlformats-officedocument.presentationml.presentation",
  txt = "text/plain",
  xls = "application/vnd.ms-excel",
  xlsx = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
  doc = "application/msword",
  docx = "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
  mp3 = "audio/mpeg",
  mp4 = "video/mp4",
  mpeg = "video/mpeg",
  ogv = "video/ogg",
  webm = "video/webm",
  avi = "video/x-msvideo",
}